+++ src/app/Components/REGISTER_YOUR_SAMPLES_HERE.cs
@@ -2,9 +2,9 @@ using Components.ComponentsSamples.BottomSheets;
 using Components.ComponentsSamples.Checkboxes;
 using Components.ComponentsSamples.Chips;
 using Components.ComponentsSamples.ContextMenus;
-using Components.ComponentsSamples.FloatingActionButtons;
 using Components.ComponentsSamples.ListItems;
 using Components.ComponentsSamples.Loading;
+using Components.ComponentsSamples.Navigation;
 using Components.ComponentsSamples.Pickers;
 using Components.ComponentsSamples.Searching;
 using Components.Resources.LocalizedStrings;
@@ -32,7 +32,7 @@ public static class REGISTER_YOUR_SAMPLES_HERE
             new(SampleType.Components, LocalizedStrings.ListItems, () => new ListItemsSamples()),
             new(SampleType.Components, LocalizedStrings.Chip, () => new ChipsSamples()),
             new(SampleType.Resources, LocalizedStrings.Icons, () => new IconsSamples()),
-            new(SampleType.Components, LocalizedStrings.FloatingActionButtons, () => new FloatingActionButtonSamples()),
+            new(SampleType.Components, LocalizedStrings.Navigation, () => new NavigationSamples()),
             new(SampleType.Components, LocalizedStrings.Loading, () => new LoadingSamples())
         };
     }