+++ src/app/Playground/VetleSamples/VetleTestPage1ViewModel.cs
@@ -1,0 +1,13 @@
+using System.Windows.Input;
+
+namespace Playground.VetleSamples;
+
+public class VetleTestPage1ViewModel
+{
+    public VetleTestPage1ViewModel()
+    {
+        Navigate = new Command(() => Shell.Current.Navigation.PushAsync(new VetleTestPage2()));
+    }
+    
+    public ICommand Navigate { get; }
+}