+++ src/library/DIPS.Mobile.UI/Components/FloatingActionButtons/FloatingNavigationButton/dotnet/FloatingNavigationButtonService.cs
@@ -1,9 +1,0 @@
-// ReSharper disable once CheckNamespace
-namespace DIPS.Mobile.UI.Components.FloatingActionButtons.FloatingNavigationButton;
-
-public partial class FloatingNavigationButtonService
-{
-    private static partial void AttachToRootWindow(FloatingNavigationButton fab)
-    {
-    }
-}