+++ src/library/DIPS.Mobile.UI/Components/Images/Android/ImageButtonHandler.cs
@@ -1,0 +1,13 @@
+using Microsoft.Maui.Platform;
+
+// ReSharper disable once CheckNamespace
+namespace DIPS.Mobile.UI.Components.Images;
+
+public partial class ImageButtonHandler
+{
+    private static partial void TrySetTintColor(ImageButtonHandler handler, ImageButton imageButton)
+    {
+        handler.PlatformView.SetColorFilter(imageButton.TintColor.ToPlatform());
+    }
+
+}