+++ src/library/DIPS.Mobile.UI/Components/Images/ImageButton.cs
@@ -1,0 +1,5 @@
+namespace DIPS.Mobile.UI.Components.Images;
+
+public partial class ImageButton : Microsoft.Maui.Controls.ImageButton
+{
+}