+++ src/library/DIPS.Mobile.UI/Components/Images/dotnet/ImageButtonHandler.cs
@@ -1,0 +1,9 @@
+namespace DIPS.Mobile.UI.Components.Images;
+
+public partial class ImageButtonHandler
+{
+    private static partial void TrySetTintColor(ImageButtonHandler handler, ImageButton imageButton)
+    {
+    }
+
+}