+++ src/library/DIPS.Mobile.UI/Effects/Touch/dotnet/TouchPlatformEffect.cs
@@ -1,0 +1,15 @@
+// ReSharper disable once CheckNamespace
+namespace DIPS.Mobile.UI.Effects.Touch;
+
+public partial class TouchPlatformEffect
+{
+    protected override partial void OnAttached()
+    {
+        
+    }
+
+    protected override partial void OnDetached()
+    {
+        
+    }
+}