+++ src/library/DIPS.Mobile.UI/Resources/Colors/ColorName.cs
@@ -1,6 +1,6 @@
 /*
 Do not edit directly,
-generated 5/26/2023 6:01:43 PM from DIPS.Mobile.DesignTokens
+generated 6/2/2023 8:12:00 AM from DIPS.Mobile.DesignTokens
 */
 
 namespace DIPS.Mobile.UI.Resources.Colors